+++ .gitignore
@@ -13,3 +13,4 @@ __pycache__/
 
 # data related files
 recommender/data/movielens/.movielens
+recommender/data/avazu_ctr/.avazu_ctr
